fix(scanning): isolate the prune test, and fall back on a bad SYNC_LOG_KEEP
Review round 1. The prune test ran against the checkout's real
storage/logs and deleted a scan log it did not create; the suite now
switches the app to a temporary storage path and removes it afterwards.
A blank or non-numeric SYNC_LOG_KEEP cast to 0, which means keep
everything, so a typo silently disabled pruning; it now falls back to
the default of 30, with a test. The variable is documented beside
SYNC_LOG_LEVEL. Formatted with mago as CI requires.

+++ config/koel.php
@@ -204,15 +204,15 @@ return [
     'sync_log_level' => env('SYNC_LOG_LEVEL', 'error'),
 
     /*
-    |--------------------------------------------------------------------------
-    | Scan log retention
-    |--------------------------------------------------------------------------
-    |
-    | How many storage/logs/sync-*.log files to keep. Each scan that has
-    | something to report writes one. 0 keeps them all.
-    |
-    */
-    'sync_log_keep' => (int) env('SYNC_LOG_KEEP', 30),
+     |--------------------------------------------------------------------------
+     | Scan log retention
+     |--------------------------------------------------------------------------
+     |
+     | How many storage/logs/sync-*.log files to keep. Each scan that has
+     | something to report writes one. 0 keeps them all.
+     |
+     */
+    'sync_log_keep' => env('SYNC_LOG_KEEP', 30),
 
     'proxy_auth' => [
         'enabled' => env('PROXY_AUTH_ENABLED', false),
